+++ src/models/persistence/CurrentIlluminationDataPoint.ts
@@ -1,9 +1,0 @@
-export class CurrentIlluminationDataPoint {
-  constructor(
-    public roomName: string,
-    public deviceID: string,
-    public currentIllumination: number,
-    public date: Date,
-    public lightIsOn: boolean,
-  ) {}
-}