+++ src/server/devices/baseDeviceInterfaces/iIlluminationSensor.ts
@@ -1,3 +1,5 @@
-export interface iIlluminationSensor {
+import { iRoomDevice } from './iRoomDevice';
+
+export interface iIlluminationSensor extends iRoomDevice {
   currentIllumination: number;
 }