+++ src/server/devices/hmIPDevices/hmIpBewegung.ts
@@ -1,6 +1,6 @@
 import { HmIPDevice } from './hmIpDevice';
 import { DeviceType } from '../deviceType';
-import { CountToday, CurrentIlluminationDataPoint, LogLevel, MotionSensorSettings } from '../../../models';
+import { CountToday, LogLevel, MotionSensorSettings } from '../../../models';
 import { Utils } from '../../services';
 import { iIlluminationSensor, iMotionSensor } from '../baseDeviceInterfaces';
 import { IoBrokerDeviceInfo } from '../IoBrokerDeviceInfo';
@@ -59,15 +59,7 @@ export class HmIpBewegung extends HmIPDevice implements iIlluminationSensor, iMo
 
   private set currentIllumination(value: number) {
     this._currentIllumination = value;
-    Utils.dbo?.persistCurrentIllumination(
-      new CurrentIlluminationDataPoint(
-        this.info.room,
-        this.info.devID,
-        value,
-        new Date(),
-        this.room?.LampenGroup?.anyLightsOn() ?? false,
-      ),
-    );
+    Utils.dbo?.persistIlluminationSensor(this);
   }
 
   public addMovementCallback(pCallback: (pValue: boolean) => void): void {